Container-Design-and-Deployment/Assignment-2/app/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.108.0
h11==0.14.0
idna==3.6
mysql-connector-python==8.2.0
protobuf==4.21.12
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
